requirements.txt
face_recognition~=1.3.0
mysql-connector-python~=8.3.0
numpy~=1.26.4
streamlit~=1.33.0
pillow~=10.2.0
opencv-python==4.8.1.78
scikit-image~=0.22.0
streamlit_js_eval~=0.1.7

